# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tres Service API Documentation",
        "description": "API documentation for Tres Service application",
        "contact": {
            "email": "sibomanaedouard974@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/trash": {
            "get": {
                "tags": [
                    "Trash & Restore"
                ],
                "summary": "List all trashed files & folders",
                "description": "Retrieves a list of all soft-deleted files and folders for the authenticated user",
                "operationId": "listTrash",
                "responses": {
                    "200": {
                        "description": "Trashed items retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trashed items retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "files": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/File"
                                                    }
                                                },
                                                "folders": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Folder"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/trash/restore": {
            "post": {
                "tags": [
                    "Trash & Restore"
                ],
                "summary": "Restore a file or folder from trash",
                "description": "Restores one or more files or folders from the trash. If restoring a folder, its contents are also restored.",
                "operationId": "restoreTrashItem",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "file_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ],
                                        "nullable": true
                                    },
                                    "folder_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ],
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Selected items restored successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Selected items restored successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "One or more items not found in trash"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/trash/file/{id}": {
            "delete": {
                "tags": [
                    "Trash & Restore"
                ],
                "summary": "Permanently delete single file from trash",
                "description": "Permanently deletes a single file that is currently in the trash",
                "operationId": "permanentlyDeleteFile",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the file in trash",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "File permanently deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "File permanently deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "File not found in trash"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/trash/file/all": {
            "delete": {
                "tags": [
                    "Trash & Restore"
                ],
                "summary": "Empty trash (all files/folders)",
                "description": "Permanently deletes all files and folders in the trash for the authenticated user",
                "operationId": "emptyTrash",
                "responses": {
                    "200": {
                        "description": "Trash emptied successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Trash emptied successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/share/{link}": {
            "get": {
                "tags": [
                    "Sharing & Links"
                ],
                "summary": "Access public share link",
                "description": "Accesses content shared via a public link. Requires password if protected.",
                "operationId": "accessPublicShareLink",
                "parameters": [
                    {
                        "name": "link",
                        "in": "path",
                        "description": "Unique share link string",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "linkpass",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Content retrieved successfully",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "401": {
                        "description": "Unauthorized: Invalid password or link expired"
                    },
                    "404": {
                        "description": "Link not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "File": {
                "title": "File",
                "description": "File model",
                "required": [
                    "id",
                    "user_id",
                    "folder_id",
                    "name",
                    "file_path",
                    "file_type",
                    "file_size",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "File ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Owner User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "folder_id": {
                        "description": "Parent Folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "File name",
                        "type": "string",
                        "example": "document.pdf"
                    },
                    "description": {
                        "description": "File description",
                        "type": "string",
                        "example": "A brief summary of the document.",
                        "nullable": true
                    },
                    "file_path": {
                        "description": "Internal path to the file storage",
                        "type": "string",
                        "example": "files/user_1/folder_1/document.pdf"
                    },
                    "file_type": {
                        "description": "MIME type of the file",
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "file_size": {
                        "description": "File size in bytes",
                        "type": "integer",
                        "example": 1024000
                    },
                    "total_downloads": {
                        "description": "Total number of times the file has been downloaded",
                        "type": "integer",
                        "example": 50
                    },
                    "last_access_date": {
                        "description": "Last access timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T14:30:00Z",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp (for soft delete)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "Folder": {
                "title": "Folder",
                "description": "Folder model",
                "required": [
                    "id",
                    "user_id",
                    "name",
                    "privacy",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "description": "Folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "user_id": {
                        "description": "Owner User ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "parent_id": {
                        "description": "Parent folder ID",
                        "type": "integer",
                        "format": "int64",
                        "example": null,
                        "nullable": true
                    },
                    "name": {
                        "description": "Folder name",
                        "type": "string",
                        "example": "My Documents"
                    },
                    "privacy": {
                        "description": "Folder privacy",
                        "type": "string",
                        "enum": [
                            "public",
                            "private"
                        ],
                        "example": "private"
                    },
                    "password_protected": {
                        "description": "Is folder password protected?",
                        "type": "boolean",
                        "example": false
                    },
                    "allow_download": {
                        "description": "Allow download from this folder",
                        "type": "boolean",
                        "example": true
                    },
                    "watermark_images": {
                        "description": "Watermark images in this folder",
                        "type": "boolean",
                        "example": false
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-08-09T12:00:00Z"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp (for soft delete)",
                        "type": "string",
                        "format": "date-time",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}